+++ @kiva/kv-shop/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.28](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.10.27...@kiva/kv-shop@1.10.28) (2024-04-29)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
+## [1.10.27](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.10.26...@kiva/kv-shop@1.10.27) (2024-04-26)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
+## [1.10.26](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.10.25...@kiva/kv-shop@1.10.26) (2024-04-26)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
+## [1.10.25](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.10.24...@kiva/kv-shop@1.10.25) (2024-04-25)
+
+**Note:** Version bump only for package @kiva/kv-shop
+
+
+
+
+
 ## [1.10.24](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-shop@1.10.23...@kiva/kv-shop@1.10.24) (2024-04-23)
 
 **Note:** Version bump only for package @kiva/kv-shop